(test) Test response body has directions for origin/dest requested
I didn't notice that the lat/lngs in the test query object previously used
were located inside buildings. Thus, they did not match the exact lat/lngs
of the starting and ending points in the directions object that is sent
back from the Google Maps API.

I changed the test query object to match exactly the starting and
ending point that the Google Maps API will use so that I can test the
server is not modifying the results and is actually sending back the
correct directions.